relocate xxHash64
