+++ llvm/unittests/Support/xxhashTest.cpp
@@ -32,18 +32,6 @@ static void fillTestBuffer(uint8_t *buffer, size_t len) {
   }
 }
 
-TEST(xxhashTest, Basic) {
-  EXPECT_EQ(0xef46db3751d8e999U, xxHash64(nullptr, 0));
-  EXPECT_EQ(0x33bf00a859c4ba3fU,
-            xxHash64(reinterpret_cast<const uint8_t *>("foo"), 3));
-  EXPECT_EQ(0x48a37c90ad27a659U,
-            xxHash64(reinterpret_cast<const uint8_t *>("bar"), 3));
-  EXPECT_EQ(0x69196c1b3af0bff9U,
-            xxHash64(reinterpret_cast<const uint8_t *>(
-                         "0123456789abcdefghijklmnopqrstuvwxyz"),
-                     36));
-}
-
 TEST(xxhashTest, xxh3) {
   constexpr size_t size = 2243;
   uint8_t a[size];
